Add files via upload
This commit adds the 'final presentation 603.pptx' file, which details our analysis of flight delays. The presentation includes:

Introduction to Flight Delays: Outlines the significance of addressing flight delays and the potential improvements in air travel efficiency and passenger satisfaction.

Data Compilation: Discusses the data gathered from the Bureau of Transportation Statistics, covering flights between 2012 and 2022, which includes 35,318 rows of detailed information on various delay factors.

Analysis Procedure: Describes the steps taken in data cleaning, analysis using Apache Spark, and the development of a machine learning model to predict flight delays.

Visualizations: Includes bar graphs and pie charts that depict the analysis results, highlighting the most common delay factors and the performance of various airports regarding delays, cancellations, and diversions.

Technology Stack: Details the use of MongoDB for data storage, Apache Spark for data processing, and the integration of machine learning libraries.

This presentation is intended to provide a comprehensive overview of our project's objectives, methodologies, results, and the technologies utilized. It serves as a vital component of our project documentation and is designed for presentation to stakeholders interested in reducing flight delays and improving airline operational efficiency.